rabbit_ct_config_schema: Remove default value for {rabbit, [{log, _}]}
This helper is designed to perform exact matching between the generated
configuration and the expected value. This does not work at all if the
schema has default values for untested configuration variables.

The correct solution would be to rewrite this helper and all testsuites
using it to do pattern matching instead. But in the meantime, work
around this design issue by removing the `{rabbit, {log, _}}`
configuration key.