Systemd: Use erlang-systemd to notify systemd of service status
Until now, we were using socat(1). It was added as a dependency to our
Debian and RPM packages. Unfortunately, the availability of this package
has not always been good. We also supported erlang-sd_notify which
relied on a UDP socket. None of them had journald support.

Now, thanks to erlang-systemd, we can use the Unix socket and provide
more fine-grained status updates if we want.

Journald support is also provided by erlang-systemd. It was added to
RabbitMQ in commit 91583a0c0e2cfa9152509e04fe08c0f833fdc178 (#2940).