Add a github actions workflow for aws peer discovery
Run the aws peer discovery aws integration suite only when the plugin
changes. Arguably it could also be run when the broker changes, but we
do have the rest of the rabbitmq_peer_discovery_aws suites that will
run in those cases. Additionally it might make sense to run this
periodically, in the even that something on the aws side changes
between changes to the plugin.