Only register the first web-dispatch listener
The problem: on certain double IPv4 and IPv6 environments, web-dispatch
will start a single listener (for management for example) but will
register two listeners in the Rabbit Mnesia table. Later, RabbitMQ
will crash when trying to stop or suspend them (rabbitmq-upgrade drain)
because only one of the two listeners exist in Ranch.

The fix: we register only the first listener, because that's the one
we used to create the Ranch ref (see rabbit_networking:ranch_ref/1).