Get rid of over-engineering (part 3)
When we get acks and delete segment files, we no longer attempt
to remove writes from the buffer. We wait for the next flush to
do this. At this point we can only have acks in the write buffer,
so when we try to write acks later on we simply write nothing
if the segment no longer exists. We do it then because deleting
segments is much more rare than not, and because keeping a few
extra acks in the write buffer does not cost us much memory (if
any: we create less garbage this way).