Keep next_seq_id when doing a clean shutdown
The new index requires it because it keeps track of unacked
differently than the old index: messages that have not been
written ever are counted in unacked as well. So the new index
must receive a ack for all messages in order to delete the
file.

This is not a problem on unclean shutdowns because in that
case the recovery process acks any non-messages when going
through the segment files.