Keep next_seq_id when doing a clean shutdown
The new index requires it because it keeps track of unacked
differently than the old index: messages that have not been
written ever are counted in unacked as well. So the new index
must receive a ack for all messages in order to delete the
file.

This is not a problem on unclean shutdowns because in that
case the recovery process acks any non-messages when going
through the segment files.

(cherry picked from commit 7d949fc6c682c1a9d2823d92d35dbdde5326a295)