Use a different PGP keyserver for OpenSSL PGP keys
Certain keyservers fail seasonally & intermittently - every few months
one stops working reliabley - so then we alternate between them. Last
time we had this problem, pgpkeys.eu was failing, and now pgpkeys.uk
seems to be failing and resulting in flaky builds:

https://github.com/rabbitmq/rabbitmq-server/runs/2663558141?check_suite_focus=true

Trying pgpkeys.eu now, and preparing to go to the Ubuntu one next if
this one proves to not be as reliable as we would like it to be.

As a different approach, I am wondering if we should store these PGP
keys somewhere instead? WDYT @dumbbell?

Thanks @pjk25 for flagging this!

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>