Policy values take precedence over stream args
We've decided that policy values such as max-length-bytes and max-age
should take precedence over arguments provided during stream
declaration. Previously the lower of the values was used which meant
that a stream retention period could not be extended beyond the initial
definition. Since streams are long lived, retention requirements can
change as the system evolve and since they contain data, deleting them
and redeclaring with different arguments not feasible. With this change,
retention can always be adjusted through a policy.