Fix recovery when terms are accidentally empty
This is a fix for an issue that occurs when shutting down
a node (via SIGTERM) while the queues and more specifically
the queue index is recovering. When that happens
rabbit_recovery_terms has already started, and when
it starts it calls dets:open_file/2 which creates an
empty recovery.dets file. After the node is down and
restarted again, the node thinks the shutdown was clean
because the recovery file is there, except it is empty
and therefore the queues have lost all their state.

This results in RabbitMQ thinking there are 0 messages
in all classic queues.

To avoid this issue, we consider a shutdown to be dirty
in the case where we have a recovery file BUT we do not
find our state in the recovery terms.

To reliably reproduce the issue this fixes:

* Start a node

* Fill it with many messages (800k is more than enough)

* Wait a little and then kill the node via Ctrl+C twice
  (to force dirty recovery next start)

* Start the node again

* While it says "Starting broker", after waiting
  about 5 seconds, send a SIGTERM (killall beam.smp)
  to shutdown the node "cleanly"

* Start the node again

* Management will show 0 messages in all classic queues