Print Erlang and SSL library on boot, right under the logo
This is purely for information purposes. We have often seen users
copy-paste console output when reporting issues, and they miss the most
important information: Erlang & SSL library versions. After this change,
we are less likely to have ask about the Erlang & SSL library versions
via follow-up questions.

emu_flavor was introduced in Erlang 24. Erlang 23 has not JIT so we can
always return "emu".

FWIW, we discourage putting new code in rabbitmq_common as this is meant
to be shared between the broker and the client. Also, keeping the
function definitions close to where they are called is a good general
practice.

We discussed the benefits of having the OS line with @dumbbell &
@gerhard and concluded that it's not worth the effort.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>
(cherry picked from commit 52bd0d0adfc5806a12135280b375304aedeba965)