Per-queue classic queue store
This currently works both with confirms and not.
It currently always writes to the per-queue store,
it would be good to write fan-out messages to the
shared store though.

It would be good to remove the usage of MsgId except
when the shared store is needed.

(cherry picked from commit b439d2e6bb4fba4a3baa7a8d404734e4d1c2f7a0)