Track delivers per-queue rather than per-message
Because queues deliver messages sequentially we do not need to
keep track of delivers per message, we just need to keep track
of the highest message that was delivered, via its seq_id().

This allows us to avoid updating the index and storing data
unnecessarily and can help simplify the code (not seen in this
WIP commit because the code was left there or commented out
for the time being).

Includes a few small bug fixes.

(cherry picked from commit 33fada88471c238bc4d4294831b1fe2cb9ab369c)