Fix line breaks and add extra info for release notes in 3.9.2
Mentioning the release-notes change is important, as we want to change
the way contributors (including us) make changes. Release notes should
be part of the change, and the reasoning behind it is in the 3.9.2
release notes.

Removing any mention of the Grafana dashboard as it is not something
that end-users will see. We didn't mention it in the 3.9.1 release
notes, and we are even thinking of removing all Grafana dashboards from
rabbitmq_prometheus. This explains why:
https://github.com/rabbitmq/release-engineering/issues/11#issuecomment-887627938

cc @michaelklishin

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>