Faster implementation of min-masters
min-masters location strategy iterates over all the existing queues to
find the perfect place for a new queue. Unfortunately, that means it
gets slower and slower as the number of queues grows. While an
individual queue creation is still relatively quick, importing
definitions or restarting a cluster can be very slow.

This is an alternative implementation that relies on the size of the queue_metrics
ETS table. It's a bit less accurate as the metrics may show up a moment
after the queue was declared and therfore, when creating a small number
of queues, they may be distributed unevenly but that's probably a minor
problem, compared to a node startup of 30-60 minutes when there are many
queues.

Accuracy may be addressed when we have a seshat metric, which should,
I assume, always provide the exact number of queues on a given node (this
will likely also improve performance further, since reading a single
value should be faster than counting ETS table rows).