Fix memory leaks
Confirms would be added to the store state and never removed
when they are not requested by the publisher.

The store cache would grow to a much larger size than intended
because the message size was used rather than something closer
to what is really stored in the memory cache. Now the entry
size on disk is used, which is a TTB of what we put in the cache.