Avoid force pushing in update-otp-for-oci workflow
The standard behavior of the action used in this automation is to
force push the branch if already exists. In this case we will just
short circuit our workflow. It will mean that the action can no longer
automatically close the PR if the diff has converged to zero, but that
seems worth the tradeoff.