Stream queue: use local pid for offset listeners
When a consumer reaches the end of a stream it need to register an
offset listener with the local stream member so that it can be notified
when new stream messages are committed. The stream queue implementation
for some reason registered offset listeners with the leader, not the local
member.

(cherry picked from commit db3944cfc4eb80f0ee261f41b4fdaf331399f186)