Handling of `stream_not_found` return value when init a stream client
This error probably shouldn't happen if the system is correct, however
`stream_not_found` is a valid return value for the functions called
- it comes from `register_listener`.
Not handling it causes a case_clause (and long stacktrace) that is sent
to the client as an internal error, instead of a known protocol error.
Seen while debugging another issue.

(cherry picked from commit 141f95598e7d42219db486c688c7a3466398c008)