Fix persistent message count display in mgmt UI
Fixes #3396

As can be seen in
https://github.com/rabbitmq/rabbitmq-server/blob/9cf18e83f279408e20430b55428a2b19156c90d7/deps/rabbitmq_management/priv/www/js/tmpl/queue.ejs#L171
there should be 3 colums for a classic queue.

While it's possible to share the same field with the quorum queues
case, I did it this way so that conditions for the labels and for the
values will have the same shape.

(cherry picked from commit a3b02bd01c79fe86dfda822d22d793fbc4e097ab)