Maintain order when dead-lettering rejected messages in quorum queues
Before this commit, when a client consumes from a quorum queue and
rejects many messages, the order in which the messages got dead-lettered
is not the same as the order in which the messages got rejected.

Classic queues already maintain the order.

(cherry picked from commit e3ccefbf39403e03c732cf0e46ba9dc7d8099a9e)