Ensure index files with holes get removed
On dirty recovery the count in the segment file was already
accurate. It was not accurate otherwise as it assumed that
all messages would be written to the index, which is not
the case in the current implementation.

(cherry picked from commit 469788a820b57d981b1c37a1f2833f841e785833)