QQ: better handle repeated requeues.
Repeated re-queues are a problem for quorums queues. A consumer
that repeatedly re-queues all messages may eventually cause the
quorum queue to run out of messages as the log cannot be truncated
if messages are not consumed in a fifo-ish order.

This change addresses this as follows:

All return commands are not send directly to the log but instead
are sent to the aux machine which evaluates how the return should be
processed. After this decision it will use the new "append" effect
to add it to the log as before.

If the queue has a delivery_limit configured the return command
will be appended as before and the message will be returned
to the front of the queue as before. This is safe as eventually
the message will be dropped or dead-letter when it reaches
it's delivery limit.

If the queue has _not_ configured a delivery_limit the return
will be turned into a new command that includes the original
message in full and will be returne to the back of the queue.
This ensure that messages in the queue will be cycled in fifo-ish
order and thus snapshots will be taken.

(cherry picked from commit 795b5b576c187861f76ba54b7e605a8788057fd5)